refactoring final grade calculator
