+++ src/pages/Home.css
@@ -87,7 +87,7 @@
   justify-content: center;
   align-items: center;
 
-  margin-top: 20vw;
+  margin-top: 20vh;
 }
 
 .trademark__text {
